Revert "Implement more memory efficient deposit merkle tree (#5501)"
This reverts commit 784799f07c9df6676dc32d7cbd198d97e19489e6.
