+++ CHANGELOG.md
@@ -18,7 +18,6 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
 - Upated ropsten testnet config to include extremely high TTD and enable proposer boost.
-- Reduced memory requirements for storing the deposit merkle tree.
 - Changed the default maximum peers count from 74 to 100 (`--p2p-peer-upper-bound 74` was old setting)
 - Update proposer boost weighting to 40%.
 - Update `BeaconBlocksByRange` to only return the first block if the step is greater than 1, in line with 1.20 spec.
