+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/DepositProvider.java
@@ -13,7 +13,6 @@
 
 package tech.pegasys.teku.validator.coordinator;
 
-import static java.util.Collections.emptyList;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ONE;
 
 import java.util.NavigableMap;
@@ -28,7 +27,6 @@ import tech.pegasys.teku.ethereum.events.SlotEventsChannel;
 import tech.pegasys.teku.ethereum.pow.api.DepositsFromBlockEvent;
 import tech.pegasys.teku.ethereum.pow.api.Eth1EventsChannel;
 import tech.pegasys.teku.ethereum.pow.api.MinGenesisTimeBlockEvent;
-import tech.pegasys.teku.ethereum.pow.merkletree.DepositTree;
 import tech.pegasys.teku.infrastructure.logging.EventLogger;
 import tech.pegasys.teku.infrastructure.metrics.TekuMetricCategory;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
@@ -43,6 +41,8 @@ import tech.pegasys.teku.spec.datastructures.operations.DepositWithIndex;
 import tech.pegasys.teku.spec.datastructures.state.Checkpoint;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.util.DepositUtil;
+import tech.pegasys.teku.spec.datastructures.util.MerkleTree;
+import tech.pegasys.teku.spec.datastructures.util.OptimizedMerkleTree;
 import tech.pegasys.teku.storage.api.FinalizedCheckpointChannel;
 import tech.pegasys.teku.storage.client.RecentChainData;
 
@@ -55,7 +55,7 @@ public class DepositProvider
 
   private final RecentChainData recentChainData;
   private final Eth1DataCache eth1DataCache;
-  private final DepositTree depositMerkleTree;
+  private final MerkleTree depositMerkleTree;
 
   private final NavigableMap<UInt64, DepositWithIndex> depositNavigableMap = new TreeMap<>();
   private final Counter depositCounter;
@@ -77,7 +77,8 @@ public class DepositProvider
     this.eth1DataCache = eth1DataCache;
     this.spec = spec;
     depositUtil = new DepositUtil(spec);
-    depositMerkleTree = new DepositTree();
+    depositMerkleTree =
+        new OptimizedMerkleTree(spec.getGenesisSpecConfig().getDepositContractTreeDepth());
     depositCounter =
         metricsSystem.createCounter(
             TekuMetricCategory.BEACON,
@@ -97,30 +98,32 @@ public class DepositProvider
               }
 
               depositNavigableMap.put(deposit.getIndex(), deposit);
-              depositMerkleTree.pushLeaf(deposit.getData().hashTreeRoot());
+              depositMerkleTree.add(deposit.getData().hashTreeRoot());
             });
     depositCounter.inc(event.getDeposits().size());
     eth1DataCache.onBlockWithDeposit(
         event.getBlockTimestamp(),
         new Eth1Data(
             depositMerkleTree.getRoot(),
-            UInt64.valueOf(depositMerkleTree.getDepositCount()),
+            UInt64.valueOf(depositMerkleTree.getNumberOfLeaves()),
             event.getBlockHash()));
   }
 
   @Override
   public void onNewFinalizedCheckpoint(
       final Checkpoint checkpoint, final boolean fromOptimisticBlock) {
-    final BeaconState finalizedState = recentChainData.getStore().getLatestFinalized().getState();
-    final UInt64 depositIndex = finalizedState.getEth1DepositIndex();
-    pruneDeposits(depositIndex);
-    synchronized (this) {
-      if (depositIndex.isGreaterThanOrEqualTo(finalizedState.getEth1Data().getDepositCount())
-          && depositMerkleTree.getDepositCount()
-              >= finalizedState.getEth1Data().getDepositCount().longValue()) {
-        depositMerkleTree.finalize(finalizedState.getEth1Data());
-      }
-    }
+    recentChainData
+        .retrieveBlockState(checkpoint.getRoot())
+        .thenAccept(
+            finalizedState -> {
+              if (finalizedState.isEmpty()) {
+                LOG.error("Finalized checkpoint state not found.");
+                return;
+              }
+              final UInt64 depositIndex = finalizedState.get().getEth1DepositIndex();
+              pruneDeposits(depositIndex);
+            })
+        .reportExceptions();
   }
 
   private synchronized void pruneDeposits(final UInt64 toIndex) {
@@ -213,22 +216,13 @@ public class DepositProvider
    * @param fromDepositIndex inclusive
    * @param toDepositIndex exclusive
    * @param eth1DepositCount number of deposits in the merkle tree according to Eth1Data in state
+   * @return
    */
   private SszList<Deposit> getDepositsWithProof(
       UInt64 fromDepositIndex, UInt64 toDepositIndex, UInt64 eth1DepositCount, long maxDeposits) {
     final AtomicReference<UInt64> expectedDepositIndex = new AtomicReference<>(fromDepositIndex);
     final SszListSchema<Deposit, ?> depositsSchema = depositsSchemaCache.get(maxDeposits);
     final SszBytes32VectorSchema<?> depositProofSchema = Deposit.SSZ_SCHEMA.getProofSchema();
-    // No deposits to include so don't bother rewinding the merkle tree.
-    if (fromDepositIndex.equals(toDepositIndex)) {
-      return depositsSchema.createFromElements(emptyList());
-    }
-    if (depositMerkleTree.getDepositCount() < eth1DepositCount.intValue()) {
-      throw MissingDepositsException.missingRange(
-          UInt64.valueOf(depositMerkleTree.getDepositCount()), eth1DepositCount);
-    }
-    final DepositTree merkleTree =
-        depositMerkleTree.getTreeAtDepositIndex(eth1DepositCount.intValue());
     return depositNavigableMap
         .subMap(fromDepositIndex, true, toDepositIndex, false)
         .values()
@@ -241,7 +235,9 @@ public class DepositProvider
               }
               expectedDepositIndex.set(deposit.getIndex().plus(ONE));
               SszBytes32Vector proof =
-                  depositProofSchema.of(merkleTree.getProof(deposit.getIndex().intValue()));
+                  depositProofSchema.of(
+                      depositMerkleTree.getProofWithViewBoundary(
+                          deposit.getIndex().intValue(), eth1DepositCount.intValue()));
               return new DepositWithIndex(proof, deposit.getData(), deposit.getIndex());
             })
         .collect(depositsSchema.collector());
