+++ ethereum/spec/build.gradle
@@ -66,7 +66,6 @@ dependencies {
     testFixturesImplementation project(':infrastructure:bls')
     testFixturesImplementation project(':infrastructure:bytes')
     testFixturesImplementation testFixtures(project(':infrastructure:bls'))
-    testFixturesImplementation project(':infrastructure:crypto')
     testFixturesImplementation project(':infrastructure:events')
     testFixturesImplementation project(':infrastructure:time')
 }
