+++ settings.gradle
@@ -11,7 +11,6 @@ include 'ethereum:executionclient'
 include 'ethereum:executionlayer'
 include 'ethereum:networks'
 include 'ethereum:pow:api'
-include 'ethereum:pow:merkletree'
 include 'ethereum:statetransition'
 include 'ethereum:weaksubjectivity'
 include 'data:dataexchange'
